Organized my timer-related classes and improved annotations for some methods; no change in functionalites
