+++ src/main/java/ca/unknown/bot/data_access/timer/TimerDAO.java
@@ -1,15 +1,13 @@
-package ca.unknown.bot.data_access;
+package ca.unknown.bot.data_access.timer;
 
-import ca.unknown.bot.interface_interactor.TimerController;
+import ca.unknown.bot.interface_interactor.timer.TimerController;
 import com.google.gson.*;
 import ca.unknown.bot.entities.timer.Pomodoro;
 import com.google.gson.internal.LinkedTreeMap;
 import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import net.dv8tion.jda.api.entities.User;
 
-import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -18,36 +16,37 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.google.gson.JsonObject;
-import org.apache.commons.lang3.ObjectUtils;
-
 public class TimerDAO {
     /**
-    * How are Pomodoro objects are stored? TimerDAO will categorize every instance by
-    * the Discord User that created it by using /timer_create.
-    * <p></p>
-    * {User1: {timer1_name: {"breakTime":2.0,"iteration":1,"workTime":3.0}, timer2_name:
-    *     {"breakTime":9.0,"iteration":100,"workTime":3.14}
-    * }, User2: ...}
-    */
+     * A class in the data access layer. This class communicates with timerController and
+     * timerPresenter to do the followings :
+     * 1) determine the status of timer_repository.json
+     * 2) based on its status, it selects a way to write to timer_repository.json
+     * 2) fetches data into a readable data structure so that timerPresenter can process
+     *
+     * Pomodoro instances are stored in the following format as a json file.
+     * {user_id: [{"name": "name_of_timer, "map": {"breakTime":5.0,"iteration":3,"workTime":25.0}}]}
+     *
+     * 1) user_id is a String object that comes from user.toString()
+     * 2) user_id is mapped onto ArrayList. It contains the information about a Pomodoro instance,
+     * which is retained as an instance of LinkedTreeMap (a custom data structure provided by GSON)
+     * 3) A LinkedTreeMap has key-value pairs. In the example above, "map" corresponds to
+     * another LinkedTreeMap.
+     *
+     * @param userAndTimer The desired User-Pomodoro map to be loaded or saved
+     * @param filename The name of the timer (Pomodoro) so users can look through the different
+     * Pomodoro instances.
+     */
 
-   /**
-    * GsonBuilder is a builder class that is used when one wants to create a Gson object with
-    * custom configurations.
-    *
-    * @param userAndTimer The desired User-Pomodoro map to be loaded or saved
-    * @param filename The name of the timer (Pomodoro) so users can look through the different
-    * Pomodoro instances.
-    */
 
     public void savePomodoro(Map userAndTimer, String filename) {
     /**
     * savePomodoro receives a map structure from TimerController and inspects timer_repository to
     * find out how to store a new Pomodoro instance.
-     *
-     * @param userAndTimer: a map from TimerController, whose key is a string-fied discord user and
-     *                    value is a new Pomodoro instance wrapped in an ArrayList
-     * @param filename: the location of the json file
+    *
+    * @param userAndTimer: a map from TimerController, whose key is a string-fied discord user and
+    *                    value is a new Pomodoro instance wrapped in an ArrayList
+    * @param filename: the location of the json file
     */
         // 1. Check timer_repository is an empty file
         // If it is, it will just ask GSON to write userAndTimer without modifications.
@@ -72,8 +71,8 @@ public class TimerDAO {
         // 2. Check if timer_repository doesn't contain that specific user
         else if (!checkUser(userAndTimer, filename)){
             Map repo = loadPomodoro(filename); // Copies the repository into a hashmap
-            String key = userAndTimer.keySet().toArray()[0].toString();
-            repo.put(key, userAndTimer.get(key));
+            String key = userAndTimer.keySet().toArray()[0].toString(); // turn user instance into a string
+            repo.put(key, userAndTimer.get(key)); // adds the new user-ArrayList<Pomodoro>
             Gson gson = new GsonBuilder().create();
             try (FileWriter writer = new FileWriter(filename)) {
                JsonWriter jsonWriter = new JsonWriter(writer);
@@ -94,7 +93,7 @@ public class TimerDAO {
             // 1: Encase a Pomodoro instance from userAndTimer.get(key) in a LinkedTreeMap
             LinkedTreeMap timer = TimerController.converttoLTM(pomodoro);
 
-            // 2: Put the new LinkedTreeMap in newPomodoros
+            // 2: Put the new LinkedTreeMap in newPomodoros, which is an ArrayList
             newPomodoros.add(timer);
 
             // 3: repo.put(key, newPomodoros);
@@ -110,7 +109,7 @@ public class TimerDAO {
         }
     }
 
-    // returns true if the designated json file is totally empty
+    // helper function: returns true if the designated json file is totally empty
     private boolean checkEmpty(String filename) {
         try (FileReader reader = new FileReader(filename)) {
             JsonReader jsonReader = new JsonReader(reader);
@@ -125,7 +124,7 @@ public class TimerDAO {
         return false;
     }
 
-    // returns true if the input user is not found in the json file
+    // helper function: returns true if the input user is not found in the json file
     private boolean checkUser(Map userAndTimer, String filename) {
         try (FileReader reader = new FileReader(filename)) {
             JsonReader jsonReader = new JsonReader(reader);
@@ -159,10 +158,10 @@ public class TimerDAO {
     }
 
     public ArrayList<LinkedTreeMap> loadTimers(String user, String filename) {
-    /**
+   /**
     * loadTimers returns an ArrayList that consists of Pomodoro instances.
-     *
-     * @param user an identifier that serves to find out what list to import
+    *
+    * @param user an identifier that serves to find out what list to import
     * @param filename the location of timer_repository
     * @return ArrayList of Pomodoro instances
     */
@@ -184,9 +183,10 @@ public class TimerDAO {
    public Map loadPomodoro(String filename) {
     /**
     * loadPomodoro is a reader method that fetches the timer-related data from timer_repository.json
-     * as a map.
-     *
-    * @param filename the location of timer_repository
+    * as a map. The map contains a string-fied user identifier and an arraylist, which consists of
+    * Pomodoro objects, as key and value respectively.
+    *
+    * @param filename the location of timer_repository (or any other json file)
     * @return A map representation of timer_repository.json
     */
        try (FileReader reader = new FileReader(filename)) {
