+++ src/main/java/ca/unknown/bot/data_access/timer/TimerSerializer.java
@@ -1,9 +1,7 @@
-package ca.unknown.bot.data_access;
+package ca.unknown.bot.data_access.timer;
 
 import ca.unknown.bot.entities.timer.Pomodoro;
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializationContext;
 import net.dv8tion.jda.api.entities.User;
