+++ src/main/java/ca/unknown/bot/interface_interactor/timer/TimerController.java
@@ -1,7 +1,7 @@
-package ca.unknown.bot.interface_interactor;
+package ca.unknown.bot.interface_interactor.timer;
 
 import ca.unknown.bot.entities.timer.Pomodoro;
-import ca.unknown.bot.data_access.TimerDAO;
+import ca.unknown.bot.data_access.timer.TimerDAO;
 import com.google.gson.internal.LinkedTreeMap;
 import net.dv8tion.jda.api.entities.User;
 
@@ -16,7 +16,7 @@ public class TimerController {
      * View (Discord) and converts to an appropriate datatype (Map) for TimerDAO to process.
      *
      * A typical map that maps a Discord user onto timer looks like this. (and this is how
-      *  a map is represented in timer_repository.json.)
+     *  a map is represented in timer_repository.json.)
      *
      * Map<String, ArrayList<LinkedTreeMap>>
      * {userID1: <timer1, timer2>, userID2: <timer3>}
@@ -30,6 +30,11 @@ public class TimerController {
      * (String) as a key and the Pomodoro instance they want to create as a value. It will be passed
      * onto the TimerDAO for checking duplicate user and Pomodoro instances.
      *
+      * @param name : the name of the new timer instance
+      * @param workTime : the duration of a work period (within an interval)
+      * @param breakTime : the duration of a break period (within an interval)
+      * @param iteration : the number of intervals for a timer instance
+      * @param user : the Discord user who used /timer_create
      */
         Pomodoro newTimer = new Pomodoro(workTime, breakTime, iteration, name);
         ArrayList<Pomodoro> value = new ArrayList<>();
@@ -41,7 +46,7 @@ public class TimerController {
     }
 
     public static LinkedTreeMap converttoLTM(Pomodoro pomodoro){
-    /**
+     /**
      * Converts the information above the given Pomodoro instance into a well-formed LinkedTreeMap
      * to ensure data consistency in timer_repository.
      *
